backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart